requirements.txt
# Core libraries
python-docx==1.1.2
lxml==5.2.1

# Template support
docxtpl==0.16.8

# Document merging / composition
docxcompose==1.4.0

# # Windows-only: for automating Word to convert .doc -> .docx
# pywin32==306

# # Optional: commercial library with full .doc support
# aspose-words==24.7.0